Clean up of test module
