+++ test/src/androidTest/kotlin/io/realm/NullabilityTests.kt
@@ -16,6 +16,7 @@
 
 package io.realm
 
+import io.realm.util.Utils
 import test.Nullability
 import kotlin.test.AfterTest
 import kotlin.test.BeforeTest
