+++ test/src/androidTest/kotlin/io/realm/PrimaryKeyTests.kt
@@ -16,8 +16,9 @@
 
 package io.realm
 
-import io.realm.test.allPrimaryKeyFieldTypes
-import io.realm.test.rType
+import io.realm.util.TypeDescriptor.allPrimaryKeyFieldTypes
+import io.realm.util.TypeDescriptor.rType
+import io.realm.util.Utils
 import test.primarykey.NoPrimaryKey
 import test.primarykey.PrimaryKeyByte
 import test.primarykey.PrimaryKeyByteNullable
