+++ test/src/macosMain/kotlin/io/realm/util/Utils.kt
@@ -15,9 +15,18 @@
  * limitations under the License.
  */
 
-package io.realm
+package io.realm.util
 
-expect object Utils {
-    fun createTempDir(): String
-    fun deleteTempDir(path: String)
+import kotlinx.cinterop.cstr
+import kotlinx.cinterop.toKString
+
+actual object Utils {
+    actual fun createTempDir(): String {
+        val mask = "${platform.Foundation.NSTemporaryDirectory()}realm_test_.XXXXXX"
+        return platform.posix.mkdtemp(mask.cstr)!!.toKString()
+    }
+
+    actual fun deleteTempDir(path: String) {
+        platform.Foundation.NSFileManager.defaultManager.removeItemAtURL(platform.Foundation.NSURL(fileURLWithPath = path), null)
+    }
 }
